backend/requirements.txt
fastapi[all]==0.103.0
pydantic==2.3.0
firebase-admin==5.0.0
passlib==1.7.4
python-jose==3.3.0
PyJWT==2.1.0
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.5.1
pre-commit==3.4.0
pyrebase4==4.7.1
sqlalchemy==2.0.23
sqlmodel==0.0.14
psycopg==3.1.14
psycopg-binary==3.1.14
psycopg-pool==3.2.0
psycopg2-binary==2.9.9
Mako==1.3.0
alembic==1.13.0
tenacity==8.2.3
